sage.combinat.partition: Move partitions_in_box here
